- split up environment into globals (common lua) and mudglobals (musicmud stuff) - add TypeNotNil type and make string.len() take it - add a test for that - add four tests which should fail but don't, in anticipation of future   work regarding loops / function return values
