+++ standalone-tests/failing/string-len-nil.lua
@@ -1,0 +1,1 @@
+local len = string.len(nil)
