+++ standalone-tests/should_fail/add-bad-return-type.lua
@@ -1,0 +1,5 @@
+local f = function()
+        return "..."
+end
+
+local c = f() + 12
