+++ standalone-tests/should_fail/call-bad-return-type.lua
@@ -1,0 +1,5 @@
+local f = function()
+        return nil
+end
+
+f()()
