+++ standalone-tests/should_fail/simple-while-loop-bounds.lua
@@ -1,0 +1,10 @@
+local func = function() end
+
+local t = {func}
+
+local i = 1
+
+while i < 5 do
+        t[i]()
+        i = i + 1
+end
