+++ standalone-tests/should_fail/using-too-many-return-values.lua
@@ -1,0 +1,5 @@
+local f = function()
+        return nil
+end
+
+local a, b = f()
